RP2040 host mode: Add initial handling of multiple non-interrupt endpoints
See also the comment at

 https://github.com/hathach/tinyusb/pull/1434#issuecomment-1246721482

. This commit introduces a new structure to save the logical information
about an endpoint that takes turn being committed to the single hardware
epx endpoint.

Compared to using the interrupt endpoints, which is the approach of the
referenced issue, this has the advantage that we can trigger a transfer to
occur *now* by writing to SIE_CTRL instead of waiting for hardware to poll
the interrupt descriptors after the next sof packet up to 1 ms in the
future.

It has the disadvantage that we don't get to setup multiple transfers in
advance, there can only be one epx transfer that is being attempted. At the
moment, this means that only one non-interrupt transfer can be outstanding
at a time from the tinyusb user's perspective.

It should be possible to implement software queuing of transfers to hide
this from the user, but this is not being attempted in this commit.